# api/openapi.yaml
openapi: "3.1.0"
info:
  title: FamilyRoots API
  version: "1.0.0"
  description: |
    ## FamilyRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
...
  schemas:
...
    Error:
      type: object
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri
          example: "https://familyroots.app/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "The request body contains invalid fields."
        instance:
          type: string
          example: "/api/v1/trees/abc/persons"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string
          example: "req_01hzab3k7x"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          example: "sex"
        code:
          type: string
          example: "INVALID_ENUM"
        message:
          type: string
          example: "Must be one of: MALE, FEMALE, OTHER, UNKNOWN"

    # ── Pagination ──────────────────────────────────────────
...
  responses:
...
    401Unauthorized:
      description: Unauthorized — missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    403Forbidden:
      description: Forbidden — valid token but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    422UnprocessableEntity:
      description: Unprocessable Entity — validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /admin/users:
    get:
      tags: [Admin]
      operationId: listUsers
      summary: List all users in the tenant
      description: |
        Returns a paginated list of all users in the caller's tenant.
        Requires `ADMIN` app role with a verified email address.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
          description: Search by name or email (case-insensitive)
        - name: app_role
          in: query
          schema:
            type: string
            enum: [ADMIN, STANDARD, AUDITOR]
        - name: verified
          in: query
          schema: { type: boolean }
          description: Filter by email verification status
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at_desc, created_at_asc, name_asc, email_asc, last_login_desc]
            default: created_at_desc
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                required: [total, items, page, page_size, total_pages]
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUser"
                  page: { type: integer }
                  page_size: { type: integer }
                  total_pages: { type: integer }
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"

    post:
      tags: [Admin]
      operationId: createUser
      summary: Create a new user and send an activation email
      description: |
        Creates a user in the tenant with a random unusable password.
        Sends an activation email containing a link to `POST /auth/reset-password`
        where the user sets their own password.

        **Side effects:**
        - Activation email sent immediately.
        - Activity log entry created: `event_type = ADMIN_CREATE`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, given_name]
              properties:
                email:
                  type: string
                  format: email
                given_name:
                  type: string
                  minLength: 1
                  maxLength: 100
                family_name:
                  type: string
                  maxLength: 100
                  default: ""
                app_role:
                  type: string
                  enum: [ADMIN, STANDARD, AUDITOR]
                  default: STANDARD
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "409":
          description: Email already exists in tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"
  /admin/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      tags: [Admin]
      operationId: updateUser
      summary: Update a user's profile, role, or verification status
      description: |
        Partial update — all body fields are optional.

        **Verification change side effects:**
        - `email_verified = true` → sends "account verified" email, logs `ADMIN_VERIFY`.
        - `email_verified = false` → sends "verification removed" email, logs `ADMIN_UNVERIFY`.

        **Activation/deactivation via this endpoint:**
        - `is_active = true` → logs `ADMIN_ACTIVATE`.
        - `is_active = false` → logs `ADMIN_DEACTIVATE`.
        - Name/role-only changes → logs `ADMIN_UPDATE`.

        **Self-protection** (when `userId` equals the caller's own ID):
        - Cannot change own `app_role` away from `ADMIN` → `400`.
        - Cannot set own `is_active = false` → `400`.
        - Cannot set own `email_verified = false` → `400`.
        - Name changes are always allowed.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                given_name:
                  type: string
                  maxLength: 100
                  nullable: true
                family_name:
                  type: string
                  maxLength: 100
                  nullable: true
                app_role:
                  type: string
                  enum: [ADMIN, STANDARD, AUDITOR]
                is_active:
                  type: boolean
                email_verified:
                  type: boolean
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: |
            Self-protection rule violated (e.g. cannot change own role or
            deactivate own account).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

    delete:
      tags: [Admin]
      operationId: deactivateUser
      summary: Deactivate (soft-delete) a user
      description: |
        Sets `is_active = false` and immediately revokes all of the user's
        active refresh tokens (forcing sign-out on all devices).

        **Side effects:**
        - "Your account has been deactivated" email sent to the user.
        - Activity log entry created: `event_type = ADMIN_DEACTIVATE`.

        **Self-protection:** Cannot deactivate your own account → `400`.
      security:
        - BearerAuth: []
      responses:
        "204":
          description: User deactivated — no body
        "400":
          description: Cannot deactivate your own account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"
  /admin/users/{userId}/verify:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Admin]
      operationId: adminVerifyUser
      summary: Manually verify a user's email address
      description: |
        Marks `email_verified = true` without requiring the user to click an
        email link.

        **Side effects:**
        - "Your account is verified" email sent with a sign-in link.
        - Activity log entry created: `event_type = ADMIN_VERIFY`.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "404":
          $ref: "#/components/responses/404NotFound"

# ============================================================
# AdminUser schema — referenced by admin endpoints above
# (add to components.schemas in a full spec refactor)
# ============================================================
# AdminUser:
#   type: object
#   required: [id, email, app_role, email_verified, is_active, created_at]
#   properties:
#     id:               { type: string, format: uuid }
#     email:            { type: string, format: email }
#     given_name:       { type: string, nullable: true }
#     family_name:      { type: string, nullable: true }
#     app_role:         { type: string, enum: [ADMIN, STANDARD, AUDITOR] }
#     email_verified:   { type: boolean }
#     is_active:        { type: boolean }
#     last_login_at:    { type: string, format: date-time, nullable: true }
#     created_at:       { type: string, format: date-time }